Add StudentPopup resolve issue : #63
