+++ page/src/pages/Home.js
@@ -29,7 +29,7 @@ const Home = (props)=>{
             // 사용자의 major (ex. 920 => 컴과, MajorDictionary.js에 정의되어 있음)
             // type이 1일 때만 setting
             // type이 0이면(교수면) default로 ""
-            token.type = String(token.userNumber)[0] == '1' ? 0 : 1;
+            token.type = String(token.userNumber)[0] === '1' ? 0 : 1;
             const majorNumber = String(token.userNumber).substring(4, 7);
             if (token.type === 0)
                 token.major = "";
